+++ src/test/java/com/zhongkexinli/micro/serv/common/thread/ThreadPoolMonitorTest3.java
@@ -4,16 +4,12 @@ import java.util.concurrent.ExecutorService;
 
 import org.junit.Test;
 
-public class ThreadPoolMonitorTest2 {
+public class ThreadPoolMonitorTest3 {
 
     @Test
     public void test() {
           ExecutorService executorService = ThreadPoolMonitor.threadPoolMonitor(5, "fileCopy"); 
-          
-          for(int i=0;i<10;i++) {
-        	  System.out.println(i);
-        	  executorService.execute(new RunA());
-          }
+           executorService.execute(new RunA());
     }
     
     
@@ -21,11 +17,10 @@ public class ThreadPoolMonitorTest2 {
         
         @Override
         public void run() {
-        	System.out.println("run aa");
-            double random = (Math.random()*10);
-            System.out.println(random);
-            
-			 if(random>5) { throw new RuntimeException("运行时异常"); }
+            System.out.println("run a");
+            String a = null;
+            a.length();
+                    
         }
         
     }